requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
orjson==3.10.7

rank-bm25==0.2.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
transformers==4.43.3
huggingface-hub==0.24.5
requests==2.32.3

# PDF & model gereksinimleri
pymupdf==1.24.9
pypdf==4.3.1
torch==2.3.1

